Rename strategy to erc20-with-balance
